% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSdata.R
\name{ECMWF_bg_gload}
\alias{ECMWF_bg_gload}
\title{Load ECMWF background grib files}
\usage{
ECMWF_bg_gload(
  file,
  analysis = NULL,
  variables = NULL,
  varnames = NULL,
  tocelsius = TRUE
)
}
\arguments{
\item{file}{the forecast file to be loaded in grib format}

\item{analysis}{optional analysis file for reading geopotential and land-sea-mask}

\item{variables}{character list of grib variables to be read, use short grib names}

\item{varnames}{names of the variables in the output, from short grib name to full names}
}
\value{
SpatialGridDataFrame containing fields given in parameter \code{variables}.
}
\description{
The functions loads ECMWF background field from a grib file.
The file is assumed to be in standard format and to have pre-defined grid size.
The function loads the given fields and builds a SpatialsGridDataFrame.
}
\examples{
ECdata <- ECMWF_bg_gload('file.grib')

}
\seealso{
\code{\link{SpatialGridDataFrame}}
}
