% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSdata.R
\name{ECMWF_bg_load}
\alias{ECMWF_bg_load}
\title{Load ECMWF background}
\usage{
ECMWF_bg_load(file, elon = seq(-40, 72.5, by = 0.1), elat = seq(73.5, 27.5,
  by = -0.1))
}
\arguments{
\item{file}{file to be loaded im netcdf format}

\item{elon}{the assumed longitude grid in the file}

\item{elat}{the assumed latitude grid in the file}
}
\value{
SpatialPixelsDataFrame containing a field \code{temperature} in degrees Celsius.
}
\description{
The functions loads ECMWF background field from a netcfd file.
The file is assumed to be in standard format and to have pre defined grid size.
The function loads the field temperature and build a gridded SpatialsPixelDataFrame.
}
\examples{
ECMWF_bg_load('file')

}
\seealso{
\code{\link{SpatialPixelsDataFrame}}
}
