\name{MOSfieldutils-package}
\alias{MOSfieldutils-package}
\alias{MOSfieldutils}
\docType{package}
\title{
\packageTitle{MOSfieldutils}
}
\description{
\packageDescription{MOSfieldutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MOSfieldutils}
\packageIndices{MOSfieldutils}
% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MOSfieldutils}
Maintainer: \packageMaintainer{MOSfieldutils}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
\code{\link[pkg:gstat-package]{gstat}}
}
\examples{
# esimerkki tahan
}
