% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSfieldutils.R
\docType{package}
\name{MOSfieldutils}
\alias{MOSfieldutils}
\alias{MOSfieldutils-package}
\title{MOSfieldutils: utilitied for gridded MOS data}
\description{
The package provides utilities to work with gridded MOS fields.
}
\section{functions}{

Descriptions still missing, see the individual help items.

The DESCRIPTION file:
\packageDESCRIPTION{MOSfieldutils}
\packageIndices{MOSfieldutils}
}

\seealso{
\code{\link[MOSplotting]{MOSplotting}} \code{\link[fastgrid]{fastgrid}}
}
\author{
\packageAuthor{MOSfieldutils}
Maintainer: \packageMaintainer{MOSfieldutils}
}
