% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSsetup.R
\name{MOSget}
\alias{MOSget}
\alias{MOSset}
\title{MOS options}
\usage{
MOSget(x = NULL, default = NULL)

MOSset(x, y)
}
\arguments{
\item{x}{The variable to fetch}

\item{default}{The value for option if not set before}
}
\description{
Get on option value with, e.g., \code{MOSget('cov.pars')} and set it by \code{MOSset('var',value)}
Or use \code{MOS.options$field} directly.
}
