% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSfieldutils.R
\name{MOSgrid}
\alias{MOSgrid}
\title{Interpolate MOS station data on a regular grid}
\usage{
MOSgrid(
  stationsfile = NULL,
  modelgridfile = NULL,
  bgfieldfile = NULL,
  stations = NULL,
  modelgrid = NULL,
  bgfield = NULL,
  trend_model = NULL,
  cov.pars = MOSget("cov.pars"),
  fitpars = FALSE,
  uselsm = FALSE,
  usereallsm = TRUE,
  seatreshold = MOSget("seatreshold"),
  usealt = TRUE,
  altlen = MOSget("altlen"),
  variable = "temperature",
  elon = MOSget("elon"),
  elat = MOSget("elat"),
  skipmiss = TRUE,
  LapseRate = 0,
  distfun = NULL
)
}
\arguments{
\item{stations}{spatial data frame for station observations}

\item{modelgrid}{spatial data frame for model grid definition}

\item{bgfield}{spatial data frame for bakground field}
}
\value{
SpatialPixelsDataFrame
}
\description{
Interpolate MOS station data on a regular grid
}
\examples{
out <- MOSgrid()

}
\seealso{
\code{\link{MOStest}}
}
