% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSdata.R
\name{MOSstation_csv_load}
\alias{MOSstation_csv_load}
\title{Load station data from CSV file}
\usage{
MOSstation_csv_load(file, elon = NULL, elat = NULL, skipmiss = TRUE,
  variable = "temperature", adddist = FALSE)
}
\arguments{
\item{file}{Station data file in CSV format}

\item{elon}{model grid longitudes, if given the stations outside the grid are cut off}

\item{elat}{model grid latitudes}

\item{skipmiss}{do we drop row with missing values for \code{variable}}

\item{variable}{the variable name used for cheking missing values}

\item{adddist}{do we add distance to sea as a new variable using \code{MOS_stations_add_dist}}
}
\description{
Load station data from CSV file
}
