% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOStests.R
\name{MOStest}
\alias{MOStest}
\title{Test MOS code}
\usage{
MOStest(...)
}
\description{
Load example data sets provided with the package, do gridding and plot results.
}
\examples{
library(MOSfieldutils)
MOStest()
}
