% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSfieldutils.R
\name{MOSvariofit}
\alias{MOSvariofit}
\title{MOS variogram fit}
\usage{
MOSvariofit(data, trend_model = temperature ~ 1, plotit = FALSE,
  vario_model = "Exp", cov.pars = NULL, sill = 0.5, nugget = 0,
  range = 1, fit.sigmasq = TRUE, fit.phi = FALSE, fit.nugget = FALSE)
}
\arguments{
\item{data}{dataset as SpatialPixelsDataFrame}

\item{trend_model}{the assumed trend model}

\item{cov.pars}{initial values for c(sigmasq, phi ,nugget)}

\item{sill}{same as sigmasq}

\item{nugget}{same as nugget}

\item{range}{same as phi}
}
\value{
variogram model with fitted parameters
}
\description{
Fits predefined (default exponential) variogram model to MOS station data.
}
\section{Warning}{

If the defaults are changed, you shoud check the code carefully.
}

\examples{
MOSvariofit(stationdata)

}
\seealso{
\code{\link{variogram}}, \code{\link{fit.variogram}}, \code{\link{vgm}}
}
