% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSplotting.R
\name{plot_MOS_field}
\alias{plot_MOS_field}
\title{Plot gridded MOS field}
\usage{
plot_MOS_field(g, layer = 1, main = "", cmin = -40, cmax = 40,
  ncolors = 100, plot.europe = TRUE, zoom = NULL, jpegfile = NULL,
  pdffile = NULL, pngfile = NULL, stations = NULL, ...)
}
\arguments{
\item{g}{SpatialPixelDataFrame to be plotted}

\item{layer}{The layer to plot}

\item{plot.europe}{add borders of Europe, the default is TRUE}

\item{zoom}{zoom factor c(minlon,maxlon,minlat,maxlat)}

\item{stations}{data frame for station locations, added as points}
}
\description{
This function is intended for plotting gridded MOS fields.
You can add station locations and border lines to the plot.
}
\examples{
plot_MOS_field(ypred,stations=stationdata)

}
