% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSdata.R
\name{sptogrib}
\alias{sptogrib}
\title{Save SpatialGridDataFrame as ECMWF grib data}
\usage{
sptogrib(g, file, variables = NULL, varnames = NULL, gribformat = 1,
  sample = "regular_ll_sfc_grib1", tokelvin = FALSE)
}
\arguments{
\item{g}{input spatial grid}

\item{file}{the output file in grib format}

\item{variables}{character list of variables to be saved, use R name}

\item{varnames}{names of the variables in the output, use short grib names}
}
\value{
This function does not return any value-
}
\description{
The functions takes loads SpatialsGridDataFrame as input and tries to save it in a
grib file.
}
\examples{
sptogib(output,file='file.grib',variables="dewpoint",varnames="2d"))

}
\seealso{
\code{\link{SpatialGridDataFrame}}
}
