% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSdata.R
\name{ECMWF_bg_load}
\alias{ECMWF_bg_load}
\title{Load ECMWF background}
\usage{
ECMWF_bg_load(
  file,
  elon = MOSget("elon"),
  elat = MOSget("elat"),
  variables = NULL,
  varnames = NULL,
  tocelsius = NULL,
  extra = TRUE
)
}
\arguments{
\item{file}{file to be loaded in netcdf format}

\item{elon}{the assumed longitude grid in the file}

\item{elat}{the assumed latitude grid in the file}
}
\value{
SpatialGridDataFrame containing fields given in parameter \code{variables}.
}
\description{
The functions loads ECMWF background field from a netcfd file.
The file is assumed to be in standard format and to have pre-defined grid size.
The function loads the given fields and builds a SpatialsGridDataFrame.
}
\examples{
ECdata <- ECMWF_bg_load('file.nc',
             variables=c('T_2M','Tmin_2M','Tmax_2M'),
             varnames=c('temperature','minimumtemperature','maximumtemperature'),
             tocelsius=c(TRUE,TRUE,TRUE))

}
\seealso{
\code{\link{SpatialGridDataFrame}}
}
